File routines moved to separate file
