+++ src/fileroutines.h
@@ -1,0 +1,9 @@
+#ifndef FILEROUTINES_H
+#define FILEROUTINES_H
+
+#include <string>
+
+std::string basename(std::string const & path);
+std::string remove_extension(const std::string & filename);
+
+#endif // FILEROUTINES_H
