+++ src/rm_reads.cpp
@@ -10,6 +10,7 @@
 #include <stdlib.h>
 #include <unordered_map>
 
+#include "fileroutines.h"
 #include "search.h"
 #include "stats.h"
 #include "seq.h"
@@ -205,33 +206,6 @@ void filter_paired_reads(std::ifstream & reads1_f, std::ifstream & reads2_f,
             }
         }
     }
-}
-
-/*! \brief Remove an extension from a filename
- *
- *  \param[in]  filename    a name of a file
- *  \return                 the specified filename without its extension
- */
-std::string remove_extension(const std::string& filename) {
-    size_t lastdot = filename.find_last_of(".");
-    if (lastdot == std::string::npos) return filename;
-    return filename.substr(0, lastdot); 
-}
-
-/*! \brief Get a filename from a path
- *
- *  \param[in]  path    a file path
- *  \return             a filename from the specified path
- */
-std::string basename(std::string const & path)
-{
-    std::string res(path);
-    size_t pos = res.find_last_of('/');
-    if (pos != std::string::npos) {
-        res.erase(0, pos);
-    }
-    res = remove_extension(res);
-    return res;
 }
 
 /*! \brief Print program parameters */
